Clean up static and dynamic HTML rendering.
Unify static and dynamic JS files into static/index.js.
The only public API is a `render(renderMode)` function.

Rename static/index.html to static/dynamic.html because it
is used only for dynamic pages.
