--- static/static-onload.js
@@ -1,23 +1,0 @@
-// Copyright 2021 Google LLC
-//
-// Use of this source code is governed by a BSD-style
-// license that can be found in the LICENSE file or at
-// https://developers.google.com/open-source/licenses/bsd
-
-var katexOptions = {
-    delimiters: [
-        {left: "$$", right: "$$", display: true},
-        {left: "\\[", right: "\\]", display: true},
-        {left: "$", right: "$", display: false},
-        {left: "\\(", right: "\\)", display: false}
-    ],
-    // Enable commands that load resources or change HTML attributes
-    // (e.g. hyperlinks): https://katex.org/docs/security.html.
-    trust: true
-};
-
-// Render LaTeX equations in prose blocks via KaTeX.
-var proseBlocks = document.querySelectorAll(".prose-block");
-Array.from(proseBlocks).map((proseBlock) =>
-    renderMathInElement(proseBlock, katexOptions)
-);
